Move the key generation and key encryption classes in same module since they are both now config time classes, instead of the encryption being a runtime dependency.
