+++ key-encryption/pom.xml
@@ -28,6 +28,11 @@
 
         <dependency>
             <groupId>com.github.nexus</groupId>
+            <artifactId>nexus-config</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.github.nexus</groupId>
             <artifactId>nexus-encryption-jnacl</artifactId>
             <scope>test</scope>
         </dependency>
