--- nexus-keygen/pom.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>nexus-keygen</artifactId>
-
-    <name>nexus-keygen</name>
-    <packaging>jar</packaging>
-
-    <parent>
-        <artifactId>nexus</artifactId>
-        <groupId>com.github.nexus</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>argon2</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.glassfish</groupId>
-            <artifactId>javax.json</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>nexus-config</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>nexus-encryption</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>nexus-encryption-jnacl</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>key-encryption</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-    </dependencies>
-
-</project>
