+++ nexus-test/src/test/resources/config1.json
@@ -22,9 +22,8 @@
                 },
                 "type": "unlocked"
             },
-            "publicKey": {
-                "value": "/+UuD63zItL1EbjxkKUljMgG8Z1w0AJ8pNOR4iq2yQc="
-            }
+            "publicKey": "/+UuD63zItL1EbjxkKUljMgG8Z1w0AJ8pNOR4iq2yQc="
+
         }
     ],
     "unixSocketFile": "/tmp/tm1.ipc"
