+++ pom.xml
@@ -196,7 +196,6 @@
         <module>nexus-config</module>
         <module>nexus-config-cli</module>
         <module>nexus-test</module>
-        <module>nexus-keygen</module>
         <module>service-locator-parent</module>
         <module>rest-server-parent</module>
         <module>nexus-encryption-parent</module>
